Be rather more careful and use size_t rather than int or int32_t or intptr_t or even uintptr_t in a number of places - and then try to get rid of warnings re singed/unsigned comparisons that ended up being introduced. This should make the code long-term robust against multi-gigabyte heaps. Well not yet robist but at least a move in the right direction!
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3672 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/arith.h
@@ -279,7 +279,7 @@ extern sem_t kara_sem1a, kara_sem1b, kara_sem1c,
        kara_sem2a, kara_sem2b, kara_sem2c;
 #endif
 
-extern int karatsuba_parallel;
+extern size_t karatsuba_parallel;
 
 //
 // Tests on my Intel i7 4770K system running Windows 7 (64-bit) I find that
